+++ src/lib/categories.ts
@@ -1,7 +1,7 @@
 import { Category } from "@prisma/client";
-import prisma from "../../prisma/client";
+import prisma from "@/db/client";
 
 export async function getAllCategories() {
-  const categories = await prisma.category.findMany();
-  return categories;
+	const categories = await prisma.category.findMany();
+	return categories;
 }
