+++ src/lib/postUploads.ts
@@ -1,5 +1,5 @@
 import { Prisma } from "@prisma/client";
-import prisma from "../../prisma/client";
+import prisma from "@/db/client";
 import path from "path";
 import fs from "fs";
 import matter from "gray-matter";
